Move test_key_loader from nodejs_conductor to conductor for reuse
