+++ conductor_api/src/lib.rs
@@ -140,6 +140,7 @@ pub mod error;
 pub mod holochain;
 pub mod interface;
 pub mod interface_impls;
+pub mod key_loaders;
 pub mod logger;
 pub mod static_file_server;
