+++ nodejs_conductor/native/src/lib.rs
@@ -9,8 +9,6 @@ extern crate holochain_conductor_api;
 extern crate holochain_core;
 extern crate holochain_core_types;
 extern crate holochain_node_test_waiter;
-extern crate holochain_dpki;
-extern crate holochain_sodium;
 
 mod config;
 pub mod js_test_conductor;
